+++ CHANGELOG.md
@@ -33,6 +33,7 @@ All notable changes to this project will be documented in this file.
 -   Players not being able to update initiative effects
 -   Active layer sometimes resetting on reload
 -   DMs no longer being able to kick themselves
+-   Some UI components not properly updating on shape reset
 
 ## [0.20.1] - 2020-05-11